+++ Codam_Core/Philosophers/utils/clean_up_utils.c
@@ -6,7 +6,7 @@
 /*   By: lilo <lilo@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/07/24 16:28:04 by lilo          #+#    #+#                 */
-/*   Updated: 2025/07/24 17:44:24 by lilo          ########   odam.nl         */
+/*   Updated: 2025/08/22 19:00:36 by lilo          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void	free_and_null(void *incoming_memory)
 
 void	clean_forks(t_whiteboard **whiteboard)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < (*whiteboard)->nbr_philosophers)
